Rewrite Promises
This is *mostly* feature-complete. Specifically, done() no longer
propagates die() when running under certain event loops, as this was
event loop dependent and thus behavior one should not rely on
(especially as a module author who is not supposed to care about the
backend in use.)

This is not yet documented, addressing that later. The relevant test
cases also still fail, the others all seem good.