Update tests to match new eventloop independent behavior
As of my rewrite, done() doesn't swallow exceptions anymore. This
behavior was bad anyway, as it depended on the chosen eventloop, and was
flagged as 'unsafe' in the documentation.

Although there was a test for this behavior, it shouldn't stay. Module
authors couldn't depend on it (they don't get to choose the backend),
and promise chains would get broken if you did this. Calling done()
twice on the promise (or once plus a then()), and dying from the
callback, would cause the other to not be invoked.

It's gone now. So here's a patch updating the tests to reflect the new
behavior.