AE: don't use postpone, use AE::io plus a pipe
When there are no events to be handled, epoll just waits until the
timeout expires. Under EV and most other event loops, the lowest
possible timeout is 1ms. postpone{} is implemented with a timer, which
doesn't count as an event, so we end up waiting 1ms every time we
postpone{} and there's no network traffic.

Let's create fake traffic instead by using a pipe. In my tests this is
100x faster in almost all cases.